Implemented other jump instructions and fixed cmp instruction. Also updated memory map to accommodate future features.

+++ cpu.h
@@ -8,7 +8,7 @@
 #include "common.h"
 
 void init_cpu(CPU *cpu);
-int load_program(CPU *cpu, uint8_t *program);
+uint16_t load_program(CPU *cpu, uint8_t *program);
 int exec_program(CPU *cpu);
 void cpu_exception(CPU *cpu, int status, Instruction inst);
